Initial Gui commit
Gui for home, join, and login are read to go.
Gui for the control pannel hasn't been worked on yet.
I'm currently having a problem with getting to the control pannel screen
and then getting back to the home screen. We can talk about it when we
work together or someone can figure it out